Service: lay groundwork for a service to handle camera operations
The service loads on boot if the app has been opened at least once. If not, the
app will start the service on first launch. The service does a check for the
camera flash feature and stops itself if the feature is not available.

Android Studio also decided the XML formatting in the manifest needed to be
tweaked during this commit. It will probably be different next time. :P