Service: attempt to make a Bound service
Try to start/stop by binding/unbinding. Toggle (ImageButton) appears to start
service on bind and stop service on unbind, but it's flaky. Also, rename
"hidden" stuff to "overlay" (calling it what it is). It's not actually hidden,
anyway. You can see the 1x1 preview in the bottom center of the display,
currently. May want to fix by wrapping in another layout.