Service: enter foreground mode if persistence is enabled
Create a notification that will be present when the service has entered
foreground mode (theoretically, preventing the system from garbage collecting
it). This will occur if the user has enabled the persistence setting. Otherwise,
no notification will be created and the service will not be put in foreground
mode.

Also in this commit, change the service member isRunning to a private member and
implement getter method isRunning() instead. Not sure why it wasn't done this
way to begin with, must've been tired.