Service: move stopService call in main activity from onDestroy to onStop
This creates the annoying side effect of disabling the torch when you open the
Settings preference activity because the main activity's onStop is called. But
onDestroy doesn't get called if you leave the app by means other than the Back
button (using Home, Recents, etc.) so the torch will stay lit regardless of the
current persistence setting state. There should be a way to get the best of both
worlds, but for now I'm choosing the lesser of two evils.