Remove MainFragment and collapse content into parent MainActivity
A Fragment adds no value to this particular app as it really only has one view
(plus the PreferenceActivity).

Note: git detected renames on the layout files because the content was moved
from fragment_main to activity_main so the similarity index was extremely high.
In actual fact, the fragment layout files were deleted and the activity layout
files were updated.