Refactor service start timing and behavior to reduce toggle lag
Start the service with the main activity so it can acquire the camera and start
the preview while everything is loading so the user can click the toggle and not
have to wait (as long) for the torch to come on. It has an unfortunate side
effect of reintroducing the startup lag that was fixed previously by removing
the pointless use of a fragment, because now it seems to be waiting for the
service or whatever the service is doing. May need to start it in another
process (probably bad)? Or completely abandon this approach and just have the
service load on boot and use a BroadcastReceiver to communicate with it? Lastly,
I believe the implementation in this commit is abusing startService calls, which
might also be remedied by switching to a BroadcastReceiver or something of that
nature.