Service: wrap some of the longer running camera tasks in AsyncTask
Choreographer was indicating there was too much work being done in the main
thread, causing frame skipping. The two tasks in the service that were taking
the longest were acquiring the camera and starting the preview. This wraps both
in their own AsyncTask threads and eliminates the lag reported in logcat. Feels
like a bit of overkill, though - may not need to do this for acquiring the
camera.