Service: revert wrapping acquireCamera in AsyncTask
I believe this was causing a race condition (rarely, but enough that I noticed).
If the Auto On setting was enabled, occasionally the camera would be null when
CameraDevice attempted to toggle. It was hard to reproduce, but I saw it the
most when coming out of the Settings activity. I haven't seen it since applying
this change.