Listen for preference changes and notify the service if applicable
This seems to work but it feels like it needs a refactor. We tell the service
about the persistence setting both when the user enables/disables it, as well as
when we toggle the torch. Should be no reason to send that information in both
places. Specifically, we should be able to remove it from the toggle intent, as
the service should already know about it. This would require notifying the
service of the settings near the beginning of the activity lifecycle (onStart?)
since the setting could be enabled on launch and we haven't seen a preference
change event yet.