Track user settings in main activity and pass them to the service on start
On a possibly unrelated note, I thought I witnessed a race condition involving
AsyncTask while testing this build, but I was unable to reproduce it and it
occurred while messing around with tweaks pretty heavily. May be ignorable, but
I'm including this warning in case I run into it later and need to circle back.