Settings: use dedicated keys for SharedPreferences
I was attempting to use one string resource for both the string that is seen
by the user as well as the key value for the SharedPreferences setting. But I
realized that when translated, if a user were to switch languages, they would
lose their settings because it would be looking for a key under the new,
translated name. Trying to utilize these strings for dual purposes is nonsense.