Suppress lint warning about layout inflation without a parent
I'm not entirely convinced this is the correct course of action. Maybe I should
provide a parent view somehow? Lint refers to the following blog post on the
topic:

http://www.doubleencore.com/2013/05/layout-inflation-as-intended/

I think in the case of this strange overlay, providing a null parent is probably
acceptable, so I'm suppressing the warning for the time being.