Service: cancel lingering AsyncTask in onDestroy() if present
This should not be necessary, but it is an attempt to fix an obscure bug that I
cannot seem to reliably reproduce where the app crashes at runtime due to a
thrown ConcurrentModificationException. I think this only occurs in test/debug
when installing different builds, but there's no harm in checking for a running
task anyway.