Introduce LaunchActivity splash screen
There is a somewhat inevitable delay on launch while we start the service and
acquire the Camera device. This shows a fullscreen message letting the user
know the app is initializing, rather than not showing anything until the
initialization is complete (which looks like lag).

This approach has its flaws. For example, we use a theme to display the
background drawable for LaunchActivity rather than a real UI. This contains a
Bitmap for the loading graphic/text, so the text is hard-coded and cannot be
localized.

We could make the time-consuming tasks asynchronous and get the user to the
main UI almost immediately, but we don't really want them there until the
initialization is complete and we're ready to handle user input anyway.